Update MainActivity.java
Hi, hopefully this is of use to you. I have made a reliable comunication work however I had to remove the input for prt and IP since the connection is now installised at startup. I would recommend adding a connect button and using that to open the connection. Hope this works.